pyproject.toml
[tool.poetry]
name = "trunkstream"
version = "0.1.0"
description = ""
authors = ["Adam Brenden <abrenden@vistaprint.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
python-multipart = "^0.0.9"
icad-tone-detection = "^0.5"
pymongo = {version = "3.11", extras = ["srv"]}
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
faster_whisper = "^1.0.1"
celery = "^5.3.6"
redis = "^5.0.3"
flower = "^2.0.1"
uvicorn = "^0.27.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
